+++ src/test/java/org/dyn4j/geometry/SliceTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010-2020 William Bittle  http://www.dyn4j.org/
+ * Copyright (c) 2010-2022 William Bittle  http://www.dyn4j.org/
  * All rights reserved.
  * 
  * Redistribution and use in source and binary forms, with or without modification, are permitted 
@@ -32,7 +32,7 @@ import org.junit.Test;
 /**
  * Test case for the {@link Slice} class.
  * @author William Bittle
- * @version 3.1.5
+ * @version 4.2.2
  * @since 3.1.5
  */
 public class SliceTest {
@@ -330,6 +330,15 @@ public class SliceTest {
 		TestCase.assertEquals(0.329, mass.getCenter().x, 1.0e-3);
 		TestCase.assertEquals(0.0, mass.getCenter().y, 1.0e-3);
 		TestCase.assertEquals(MassType.NORMAL, mass.getType());
+		
+		mass = s.createMass(0);
+		TestCase.assertEquals(0.000, mass.getMass(), 1e-3);
+		TestCase.assertEquals(0.000, mass.getInertia(), 1e-3);
+		TestCase.assertEquals(0.000, mass.getInverseMass(), 1e-3);
+		TestCase.assertEquals(0.000, mass.getInverseInertia(), 1e-3);
+		TestCase.assertEquals(0.329, mass.getCenter().x, 1e-3);
+		TestCase.assertEquals(0.000, mass.getCenter().y, 1e-3);
+		TestCase.assertEquals(MassType.INFINITE, mass.getType());
 	}
 	
 	/**